Rename Effects namespace to Magic
Another change to make a distinction between spells and other means of
carrying magical effects. System will be collectively known as "magic
system" moving forwards (or "magic and effect system" if you're not into
the whole brevity thing).
Added new Effects folder under Magic folder to store .cs files for
BaseMagicEffect implementations.
Added empty simple test effect script just to ensure folder is added to
git. Will build this out soon.
