+++ Assets/Game/Addons/UnityConsole/Console/Scripts/DefaultCommands.cs
@@ -13,7 +13,7 @@ using DaggerfallWorkshop.Utility;
 using DaggerfallWorkshop.Game.Items;
 using DaggerfallWorkshop.Game.Weather;
 using DaggerfallWorkshop.Game.Questing;
-using DaggerfallWorkshop.Game.Effects;
+using DaggerfallWorkshop.Game.Magic;
 using DaggerfallWorkshop.Game.UserInterface;
 
 namespace Wenzil.Console
