+++ Assets/Scripts/Game/Magic.meta
