+++ Assets/Scripts/Game/Magic/Effects/TestEffect.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3233cb53451133648a74260d362181db
+timeCreated: 1520029583
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
